# improves privileges matching by using a map (O(1) instead of O(n))