--- a/CHANGES.txt
+++ b/CHANGES.txt
@@ -27,6 +27,8 @@ Changes
 Fixes
 =====
 
+ - Improved user privileges matching to constant time complexity.
+
  - Improved the error message if an invalid table column reference is used in
    ``INSERT`` statements.