--- a/enterprise/users/src/main/java/io/crate/operation/user/Privileges.java
+++ b/enterprise/users/src/main/java/io/crate/operation/user/Privileges.java
@@ -21,7 +21,6 @@ package io.crate.operation.user;
 import com.google.common.annotations.VisibleForTesting;
 import io.crate.analyze.user.Privilege;
 import io.crate.exceptions.MissingPrivilegeException;
-import io.crate.metadata.TableIdent;
 import io.crate.metadata.information.InformationSchemaInfo;
 
 import javax.annotation.Nullable;
@@ -74,7 +73,7 @@ class Privileges {
         assert ident != null : "ident must not be null if privilege class is not 'CLUSTER'";
         String schemaName;
         if (Privilege.Clazz.TABLE.equals(clazz)) {
-            schemaName = TableIdent.fromIndexName(ident).schema();
+            schemaName = Privilege.schemaNameFromTableIdent(ident);
         } else {
             schemaName = ident;
         }