… enterprise/users/src/test/java/io/crate/metadata/UsersPrivilegesMetaDataTest.java
… sql/src/test/java/io/crate/analyze/user/PrivilegeTest.java
… sql/src/test/java/io/crate/operation/user/UserPrivilegesTest.java